сделал шаблон view фильтров динамическим
