+++ src/main.js
@@ -1,26 +1,31 @@
 import { render } from './framework/render.js';
 import TripMainInfo from './view/trip-main-info-view.js';
-import TripMainControl from './view/trip-main-control-view.js';
+import TripMainFilter from './view/trip-main-filter-view.js';
 import NewPointButton from './view/trip-main-new-point-button-view.js';
 import TripListPresenter from './presenter/presenter-trip-list.js';
 import TripPointsModel from './models/model-trip-point.js';
 import TripSortForm from './view/trip-sort-form-view.js';
 import TripList from './view/trip-list-view.js';
+import { generateMockFilter } from './mock/mock-filter.js';
 
 
 const tripMain = document.querySelector('.trip-main');
 const tripEvents = document.querySelector('.trip-events');
 const tripMainInfo = new TripMainInfo();
-const tripMainControl = new TripMainControl();
-const newPointButton = new NewPointButton();
-
-const tripSortForm = new TripSortForm();
-const tripList = new TripList();
-
 
 /* Создаем экземпляр класса TripPointModel, который может вернуть с помощью метода getTripPoint()
 массив моковых данных точек маршрута */
 const tripPointsModel = new TripPointsModel();
+
+/* заполняем view фильтра данными о количестве точек, попадающих под каждый тип фильтра и типом фильтра */
+const filterPoints = generateMockFilter(tripPointsModel.getTripPoint());
+const tripMainFilter = new TripMainFilter(filterPoints);
+
+
+const newPointButton = new NewPointButton();
+const tripSortForm = new TripSortForm();
+const tripList = new TripList();
+
 
 const tripListPresenter = new TripListPresenter({
   tripEventsContainer:tripEvents,
@@ -33,7 +38,7 @@ const tripListPresenter = new TripListPresenter({
 });
 
 render(tripMainInfo, tripMain);
-render(tripMainControl, tripMain);
+render(tripMainFilter, tripMain);
 render(newPointButton, tripMain);
 
 tripListPresenter.init();
