+++ src/utils/filter.js
@@ -5,6 +5,7 @@ const filter = {
   [TimeFilter.PAST]: (points) => points.filter((point) => isPointExpired(point.dateTo)),
   [TimeFilter.PRESENT]: (points) => points.filter((point) => isActual(point.dateFrom, point.dateTo)),
   [TimeFilter.FUTURE]: (points) => points.filter((point) => isFuter(point.dateFrom)),
+  [TimeFilter.EVERYTHING]: (points) => points
 };
 
 export {filter};
