+++ src/utils/point.js
@@ -6,19 +6,19 @@ import dayjs from 'dayjs';
   если дата окончания уже прошла, то вернет true (значит точка уже прошла по времени)
 */
 function isPointExpired(dateTo){
-  return dateTo && dayjs().isAfter(dateTo, 'minute');
+  return dateTo && dayjs().isAfter(dateTo, 'milliseconds');
 }
 
 /* Функция, определяющая актуальна ли точка, если дата начала меньше или равна текущей дате и дата окончания больше или
   равна текущей дате- вернет true(то есть точка еще актуальна)
 */
 function isActual(dateFrom, dateTo){
-  return dateTo && dayjs().isSameOrBefore(dateFrom, 'minute') && dayjs().isSameOrAfter(dateTo, 'minute');
+  return dateTo && (dayjs().isSame(dateFrom, 'minute') || dayjs().isAfter(dateFrom, 'minute')) && (dayjs().isSame(dateTo, 'minute') || dayjs().isBefore(dateTo, 'minute'));
 }
 
 /* Функция проверяет является ли дата начала точки позже текущей даты, если это так - вернет true */
 function isFuter(dateFrom){
-  return dateFrom && dayjs().isAfter(dateFrom, 'minute');
+  return dateFrom && dayjs().isBefore(dateFrom, 'minute');
 }
 
 export {isPointExpired, isActual, isFuter};
